// src/images-groups/types.ts
export type ResponseImagesGroup = {
    images_group_id: string;
    name: string;
    description: string;
    image_id_1: string;
    image_id_2: string;
    image_id_3: string;
    created_at: string;
    updated_at: string;
    product: boolean;
};

export type ResultImagesGroup = {
    imagesGroupId: string;
    name: string;
    description: string;
    imageId1: string;
    imageId2: string;
    imageId3: string;
    createdAt: string;
    updatedAt: string;
    product: boolean;
};